+++ include/swares/scan_os.h
@@ -72,6 +72,9 @@ namespace bscan {
     std::string timez();
     std::string status();
     std::string kernel();
+    int64_t users();
+    int64_t processes();
+    int64_t maxProcessMemory();
     [[nodiscard]] bool is32bit() const;
     [[nodiscard]] bool is64bit() const;
     [[nodiscard]] bool isBigEndian() const;
@@ -120,6 +123,9 @@ namespace bscan {
     static std::string getTimeZone();
     static std::string getStatus();
     static std::string getKernel();
+    static int64_t getUsers();
+    static int64_t getProcesses();
+    static int64_t getMaxProcessMemory();
     static bool getIs32bit();
     static bool getIs64bit();
     static bool getIsBigEndian();
@@ -169,6 +175,9 @@ namespace bscan {
     std::string _timez;
     std::string _status;
     std::string _kernel;
+    int64_t _users;
+    int64_t _processes;
+    int64_t _maxProcessMemory;
     bool _32bit = false;
     bool _64bit = false;
     bool _bigEndian = false;